requirements.txt.txt
pandas
requests
python-dotenv
openpyxl
snowflake-connector-python
schedule